test: add tree material tracking in test case
